update delete with an error

+++ posts/delete.php
@@ -5,12 +5,23 @@
     if(isset($_GET['del_id'])){
         $id = $_GET['del_id'];
 
-        $delete = $conn->prepare("DELETE FROM posts WHERE id = :id");
-        $delete->execute([
-            ':id' => $id
-        ]);
+        $select = $conn->query("SELECT * FROM posts WHERE id =' $id'");
+        $select->execute();
+        $posts = $select->fetch(PDO::FETCH_OBJ);
 
-        header("location: http://localhost/Blog/index.php");
+        if($_SESSION['user_id'] !== $posts->user_id){
+            //header("location: http://localhost/Blog/index.php");
+        }else{
+
+            unlink("images/". $posts->img. "");
+
+            $delete = $conn->prepare("DELETE FROM posts WHERE id = :id");
+            $delete->execute([
+                ':id' => $id
+            ]);
+        } 
+
+        //header("location: http://localhost/Blog/index.php");
     }
 
 ?>
